Replace …-reset-editor-text fields with native ….defaultValue
This is a native browser feature that exists for a very long time and
is supported by virtually all browsers, as far as I'm aware of. See
https://developer.mozilla.org/de/docs/Web/API/HTMLTextAreaElement

It's even possible to dynamically change the defaultValue, but it doesn't
look like we need this. I did not used this in the test, but:
* Set the default value first by appending the text to the textarea node.
* Edit the value once by calling ….val( … ).

This gets rid of 1/4 of the HTML we transfer to the client.

Since the …[0] array access might be undefined and make the following
code fail, I'm adding ….length safety checks, just to make sure this
code never crashes. This does not change the behavior. Before, jQuery
would do the same and not fail when no element was found.

Change-Id: I6535c7a3ce6f0574cff635198281bd962b056d20