Rearrange beta features check
The beta feature flag only means anything when the BetaFeatures
extension is available, so these tests should be combined.

The motivation is just to clean up, but note that there is a small
functional change, which I think is more correct.  On any wikis where
BetaFeatures was disabled but the configuration was set to run
TwoColConflict in beta mode, we would have shown the beta feature
text on conflict pages.  Instead, we should show the default feature
text.

Bug: T249659
Change-Id: Id393f211273f1079c4b5308c2471671d59c27dfc