More narrow dependency injection in HtmlEditableTextComponent and related
By using the most narrow interface that's actually needed by a class we
reduce the amount of dependencies to other code. The more we do this, the
more these classes become their own, independent units.

As of now, not all classes can be restructured to be able to have pure
unit tests for them. Any code that reaches out to call a static method
somewhere potentially breaks this. However, this patch tries to move the
code in a direction where it depends much less on MediaWiki core and it's
database. Most notably: we don't use an actual Language object any more.

Change-Id: I3908581dbb0dccd26176d0a126d8cf5e281fc127