Use actual dependency injection for HtmlEditableTextComponent
Instead of injecting all the properties that are exclusively needed by
the HtmlEditableTextComponent class, we can inject the instance. This
will make is possible to test the two independently.

As of now this is still a very small change that does not really change
anything. The benefit will become more obvious in the following patches
(e.g. I3908581).

Change-Id: Ibf0b112b56250556771c6b6efc154ab27a82d195