Split HtmlEditableTextComponent from HtmlSplitConflictView test
Before, the …ViewTest was testing parts of the …TextComponent class.
Most notably the newline handling.

With this patch I'm mocking the …TextComponent, which means the
…ViewTest is a more pure unit test for only this class.

The same time I'm introducing more specific tests for the
…TextComponent class to properly cover a series of specific features
we care about the most (e.g. the newline tracking, that proper HTML
escaping is done everywhere, and so on).

Change-Id: I553ba4c3cc601e197a181e9725735ffb4f4ba7ba