Make sure TwoColConflictContextTest runs with Beta enabled
The CI setup does not load the Beta extension when running a pure unit
test. This patch turns this (but only this) test back into an integration
test to fix this.

Note that it's still possible to run this test without the Beta
extension, e.g. locally. That's why it's still helpful to have code in
there that makes the tests succeed (or skip) without the extension.

I marked at least one test as skipped in this case to make it visible
that the Beta extension is missing.

Change-Id: Ie3d140e1eff4b5952f61cc724dba58cf2b7f8c67