Make sure TwoColConflictHooksTest runs with BetaFeatures enabled
Same as Ie3d140e. Marking a test as a "MediaWikiUnitTestCase" means no
extension will be loaded. But this test is testing features that depend on
the BetaFeatures extension. Actually, it's able to do both. But it's much
more relevant for us to run these tests with the BetaFeatures extension
enabled because that's the only code path that relevant on our Wikimedia
cluster.

Global access in a MediaWikiIntegrationTestCase based test needs to be
wrapped in setMwGlobals() in order to benefit from base test class hygiene.

This patch also fixes an incomplete test setup in
HtmlTalkPageResolutionViewTest. \OOUI\Theme::setSingleton() was null when
the test started, and should be null when it ends. However, this shouldn't
have any effect.

This partially reverts I582632f.

Change-Id: Ifcbed8b1e63cc4015bd8eb03348afad83bb92302