Add note on core conflict screen about new UI
This will show a note to users that see the core conflict screen to
point them to the setting to enable TwoColConflict.

The note will be added to the editFormTextTop property that will be
adding the HTML to the view right underneath the edit notes that
include the explanatory text on edit conflicts.

Using this public property feels a bit like technical debt but since
the whole code will be removed eventually I think it's reasonable
regarding the alternatives.

Note that the key for the user setting has to start with the userjs
prefix otherwise it can not be saved from JS.

Keyboard interaction with the X and the checkbox is complicated. I
left that aspect out for the moment. We can try to fix that later.

Test will follow.

Bug: T246103
Change-Id: I66a4647203fc1a9e86b9f49826a89ad69a9c45ed