Separate tracking of whitespace-only paragraphs
The fact that the user can empty a block, and that block is then removed
along with all it's extra linefeeds before and after is an intentional
feature. However, this was also happening for existing whitespace-only
blocks that have not been touched by the user.

This patch introduces separate tracking for such blocks. It uses the
existing "extra linefeeds" fields for this. Now, 3 types of values are
possible:
* "2" counts extra linefeeds (on top of the 1 linefeed that's always
  added) *after* the block.
* "0,2" counts extra linefeeds *before*. This typically only happens in
  the first block, because there is no previous block that could track
  these linefeeds.
* "2,was-empty" marks blocks that contain nothing but linefeeds and should
  never be removed. There is no point in splitting the number into "after"
  and "before" in this case, which is why the field can be reused for a
  string marker.
* "0" is neither of the above. The hidden field can as well be missing
  then (which is what If7dbdd8 suggests to do).

The special case for the first block is not needed any more and can be
removed.

Bug: T250262
Change-Id: Iedb9a0cd37f41c18f47406e736207b4808567830