Clean up beta mode tests
Clearly distinguish between tests that depend on the BetaFeatures
extension and those that do not, rather than dynamically changing
test expectations.  This makes it possible to see what is actually
tested in each situation.

TODO: In later work, we should inject the ExtensionRegistry so that
it's possible to test no-BetaFeatures code paths under CI.

Change-Id: I4ebaf7dcdfce3351d3393a7dba2e2e83c76fc299