Fix minor merge issue in SplitConflictUtils::mergeTextLines
My original motivation was to add test cases for a few minor, currently
untested methods (that's why this is tagged with T246103). But I
discovered a bug:

Since we are normalizing line endings *after* merging, this can happen:
"a\r" and "b" are merged to "a\r\nb", and then normalized to "a\nb".

Expected: "a\r" is normalized to "a\n", and then merged to "a\n\nb".

Solution: Either normalize first, or merge using "\r\n". This way no
conflict between the 3 types of line endings can happen. Performance
should be ok. Would probably be worse if we would normalize all lines
individually.

Bug: T246103
Change-Id: Ib1b2fa1f0035038e03e79f022fe5e636b4b5abe7