Allow textarea scrolling & resizing in no-JS mode
I realized that the fadeout effect does not make sense in no-JS mode.
It's meant to visualize the fact that the text is cut of, *instead* of
a scrollbar. Now that we have a scrollbar it's not needed any more. On
top of that is a technical reason: the fadeout hides parts of the
scrollbar and the resize handle. It's technically impossible to have
both.

I had to do a minor change to the "number of rows" calculation as well
to avoid extremely long textareas that don't fit on one screen any more.
I decided to use the largest threshold of 18, twice. Some text is still
longer than that, but can be scrolled and resized now, whatever the
user prefers.

Note I'm intentionally not fiddling with the position of the scrollbars
in this patch. This is already done in another patch, see I699289c.

Bug: T250397
Change-Id: I78c7003d564aa95456be20baab63e2e51ff916fa