Temporarily disable new UI on mobile
I found multiple ways to possibly detect mobile:
* Checking what the users "skin" option says might not be sufficent,
  because this is not the only way the mobile mode can be enabled.
* Checking OutputPage::getTarget() === 'mobile' would be perfect, but
  we don't always have access to the current OutputPage object.
* MobileContext appears to be the way to go. Several other codebases
  do this as well.

Some codebases use MobileContext::singleton(), but it seems this is
deprecated and the service registry should be used instead.

Note this is intentionally not covered by a test. Such a test would
have a pretty horrible setup and not prove much, if anything.

Bug: T249817
Change-Id: Ie879a3ef90e7a5b317c536be807af383c786b773