Merge talk module into generic CSS ResourceLoader module
As argued in T246103#6107463 it's relevant to carefully consider if a
ResourceLoader module is worth being a separate thing. The size of
*every* page load increases just because a module *exists*, no matter
if it is loaded or not.

The talk CSS got split for three reasons:
* For cleaner separation of concerns for us developers. This is not
  changed by this patch.
* To reduce page size in non-talk situations.
* Because there is no easy way to style the `.editOptions` area in the
  talk situation only.

The talk CSS is about 0.6 KB when gzipped. This is only loaded when a
user runs into an edit conflict (see
SplitTwoColConflictHelper::getEditFormHtmlAfterContent). This is *rare*,
compared to the millions of page views that don't run into an edit
conflict. I believe we save more bytes when the module does *not*
exist.

More details:
* Hiding the "wpPreviewWidget" is already done in the parent .less file.
* I checked, and all class names mentioned in the talk styles are
  exclusive to the talk use case. Except for
  "mw-twocolconflict-split-copy", which is why this is the only one I
  wrapped to make it more specific.
* The `~` selector allows to conditionally style the `.editOptions`
  area depending on a previous element on the same level. See
  https://developer.mozilla.org/en-US/docs/Web/CSS/General_sibling_combinator

Bug: T233280
Change-Id: If0f9b016d301ff89fe4a4d2d3eeeace186759005