Merge core CSS into generic CSS RecourceLoader module
As argued in T246103#6107463 it's essential to carefully consider if
a piece of code needs to be a separate ResourceLoader module. The size
of *every* page load increases just because a module exists, no matter
if it is loaded or not.

The core CSS module is about 0.3 KB when gzipped. It's only loaded for
a tiny fraction of users: users that disabled the new UI *and* never
dismissed the message. And we only want to do this *temporarily*!

When combined, the CSS module is about 2.5 KB, instead of 2.2 KB. This
does have 2 effects:
* The user group described above loads much more CSS that's mot needed.
  But since this group is tiny, it should not matter.
* Users running into a conflict will load 0.3 KB more. This group is
  also tiny, compared to millions of page views.

I believe the total number of bytes transfered we save is bigger after
this change.

Bug: T233280
Bug: T246103
Change-Id: Ic03bd45ff81ba6a575193634825bd244974b4bf4