Merge ….CoreHint.js module into the ….init.js module
As argued in T246103#6107463 the simple fact this module exists
increases the size for every page view. There are millions, compared
to a tiny, tiny fraction of users who will ever see this "core hint".

This patch merges the code into the existing "init" module. This does
have an effect for all user groups:
* Users that see the core hint will download much more JavaScript, but
  only a tiny fraction is executed. This is acceptable because this user
  group is so tiny. Also note this is meant to be temporary anyway.
* Users that have the new UI disabled and dismissed the hint already will
  not load any core. Note there are now 2 places in the code where the
  same module is loaded, for different reasons. Neither is executed in
  this case.
* Users that run into a conflict and use the new UI will:
  * … download this little extra piece of JavaScript code. This should be
    acceptable because this user group is small as well.
  * … one extra selector will be executed every time. Again, this should
    be acceptable, because it doesn't happen often.
* In all other situations we save the bytes a registered ResourceLoader
  module needs.

I tested this in all situations I can think of, and it seems to work
fine.

Bug: T233280
Bug: T246103
Change-Id: Ibf4633cf5564b423176f7acabc7d54b3e7749594