Fix browser tests skipping on user setting
The wrongly used "this" made wdio jump to the beta pref page where
the non beta setting could never be found. As a result the test was
always skipped.

While fixing this I realized, that there always will be either a
beta or a non beta way to disable the interface so it makes sens
to try both and test both.

Change-Id: I209ce93d502d7fd13c015a9b81034b88dd740629