Mark pulsating blue dots with `cursor: pointer`
I checked, and no, I think this style does not belong to the upstream
"mw-pulsating-dot" class. In the upstream code there is no guarantee
that all blue dots are clickable. We should only mark them as clickable
if they really are, which is something that happens here in this
codebase.

The change in the .js file is pure refactoring with no effect to the
function.

I don't think this is worth a ticket, as it is probably worth 0.5
points or so.

Change-Id: I5f3f09f4fe2065272881732e43ae68e53b7a5b8c