Re-arrange all classes into meaningful namespaces
My first step was to remove everything from the mostly meaningless
"SplitTwoColConflict" namespace. Originally, this was introduced
because we kept an old version in another namespace. This is long gone.
All we have now is a single "two column" UI. Repeating this two times
in the namespace should be redundant.

In the next step I arranged classes into logical groups, following a
few existing naming schemes:
* A "Hooks" namespace is very common.
* "Html" is something we already use in e.g. FileImporter.
* "Logging" contains code that is exclusively used for logging.
* "TalkPageConflict" is a name we already use in the modules folder.

I did not bothered arranging the tests the same way. I don't think
this is necessary. It's not even possible because we are forced to use
the "unit" folder.

To do: Possibly re-arrange the files in the modules folder in a
similar way, but preferrably in a later patch.

Change-Id: I81ffebedfbb89c493fd1cbe80446d69bf7b13acd