Pull context dependency 1 level up, out of ResolutionSuggester
This removes all external dependencies from ResolutionSuggester and
turns it more into an actual "unit". This is visible in the test setup.

There is a call to a static SplitConflictUtils::…() method. But this is
really just that, a utility function, not different from e.g. substr().

This is similar to what I54a571e already did. Both are extracted from
I483db53.

Change-Id: I7b610fccca00ab2f431e87a443e90696eae50730