Mark EventLogging code with @codeCoverageIgnore
At the moment, these functions are reported to be the biggest project
risk. But that's not really true. These are no user facing features.
It's just logging. Personally, I think it is of not much value to
cover these methods with tests. The test setup would be heavy, and the
result won't tell us much. I suggest to accept it as it is, and mark it
as such, to make the coverage report more meaningful.

The alternative is to write tests. If you feel like doing so, please go
ahead.

Change-Id: I1d95b14b5c5d36cd2aff71f23d2c123387d9991c