Remove meaningless "SplitTwoColConflict" namespace
This namespace was introduced when we had 2 different UIs in 2
different namespaces. The previous UI is gone. This namespace does
not make any sense any more. Let's get rid of this legacy and start
fresh.

This is only meant as a first step. Later patches will move the
classes into more meaningful namespaces.

Change-Id: I415ddcc03a5eb04669521424309a245251bf62ac