Test literal wikitext
This is a workaround for potential test flakiness caused by other
extensions manipulating content inside of the parser output div, for
example by injecting QuickSurveys.  It's also a nice enhancement to
our tests because it covers the end-to-end integration for HTML tags
in wikitext, and is sensitive to whitespace, which has already been
the source of several subtle bugs.

Bug: T254974
Change-Id: I9e3970731d88527f213f382ba7e2b6bb405c7ff3