Provide a page to copy the submitted text
Introducing a cache that saves the submitted text when entering an
edit conflict without JS and a special page where this text can be
re-visited to further work with it instead of using the two column
UI.

Changed the LinkTarget typing to Title to have access to
getPrefixedDbKey.

Test will be added in follow ups.

Bug: T254440
Change-Id: I9285b72e9a9f5c95f691850aca13f1e0016e5cc7