Fix possible cache key conflict in SubmittedTextCache
The user ID and session ID use the same spot in the cache key. Which
means whenever a user ID is generated that looks like a user ID, one
user can see the text from another user.

Introduced in I6554fb1. To be fair, an actual conflict scenarion is
extremely unlikely, if even possible. Still worth fixing, in my
opinion.

This patch also updates the code to accept WebRequest::getSessionId()
returning null. At least when we don't need the session anyway, for
registered users. Fails hard for anonymous users. Note this is not
different from the previous implementation, which also failed hard
(with a runtime error). It's just a better message now.

Bug: T254440
Change-Id: I7ded84ea79487c18017f9530c993e22971cd93b1