Use same "publishchanges" button label in all messages
"publishpage" was introduced in Iac967ed. But this is for new pages. The
pages we are dealing with here are never new (well, they should not be).

Strictly speaking there is a $wgEditSubmitButtonLabelPublish config flag
to turn the older "savechanges" button label into "publishchanges". See
EditPage::getSubmitButtonLabel(). This flag is false by default, but true
on the entire Wikimedia cluster.

An older patch used "savechanges", see If7ac75b (T248556).

I suggest to hard-code "publishchanges" for now because this is what's
more relevant for us.

Bug: T250501
Change-Id: Ia53891ef8b7aeb7d1539fe8b948a11313847b983