Restore missing header colors in no-JS mode
This is about the colors in the two headers that say "Latest revision as
of 4 minutes ago" and "Your revision". The colors have been removed in
Ib2b332f years ago. This was done because of the old interface, because
it would not show any color in no-JS mode. This is not true for the
current interface any more. The two columns and the radio buttons are
yellow/blue even in no-JS mode. Not showing the colors is inconsistent
and confusing.

Change-Id: I789ecc024cb0ce0bf56cef754bd658332065f90a