Change "entry" to "comment" in talk page conflict resolution
I found the word "entry" very weird.

I understand that it was worded this way because the thing you wrote
on the talk page could be e.g. a vote rather than a real comment.

Even so, I think that calling it a comment would be more intuitive
and less distracting.

Bug: T261852
Change-Id: I2ff26d9330ce86fe9fb17c0778bf3b317a413e05