Increase browser test stability
This patch does two things to increase the general stability of the
browser tests.

First is increasing the timeout on some of the waits. I saw a test
failing there even though the elements behaved like they should.
1000ms seem to be a bit too few.

Additionally the patch switches to the usage of the API for setting
and getting the preferences where possible.

Bug: T266711
Change-Id: I865aee8a8c7ec3c6327f4a60eb8689628f9544a9