Fix empty lines being skipped during merge
The diff algorithm returns the empty space between two changed
paragraphs as an unchanged "copy" line. This shows up in the
UI as an empty gray box. Later the merge algorithm acts as if
the user emptied this box. In this special case the line is
entirely skipped.

In other words: The merge algorithm assumes there are no empty
boxes in the UI. If there is an empty box, it's because the user
emptied it. This was not always true.

This change gets rid of empty boxes, merging them into the
previous non-empty box.

This is done for both talk as well as non-talk use-cases. The
tests should proof nothing breaks.

Bug: T250402
Bug: T266860
Change-Id: I8ca080327aead4929fb31cee629d6d8afaf89776