Re-add check for opt out user setting
This was accidentally changed in a way that it does not work as it
should in I865aee8a8c7ec3c6327f4a60eb8689628f9544a9.

The patch re-adds that check but using a more stable appoach checking
the beta settings. When these are present, the user setting should be
absent.

Change-Id: I717f0c263155bf383d7e3a02edd9a0821acac55f