Use actual Title objects instead of mocks where possible
In most of these cases the only thing that is ever used is
the namespace. This never triggers database calls, as far
as I can see.

The main advantage of this change is that the test does not
encode implementation details any more (i.e. which Title
methods are called).

Change-Id: I6b15e997c5b66eaaa3c5a46df77e316bfdafa412