Fix two edge-case bugs in whitespace-only detection
The same input causes both bugs: When a text starts with
newlines, but these newlines are not changed, the diff
starts with a "copy" block that appears empty (contains only
newlines). The detection algorithm:
* Tries to access the previous element, but there is none.
  Just because it tries, it accidentially creates an element
  `-1 => null`.
* Appends the leading newlines instead of prepending them.

I believe neither of these issues is bad enough to be worth
a backport. It's extremely rare anyway.

Bug: T266860
Bug: T268313
Bug: T268314
Change-Id: Id7ee2aa3012989ff8eaa48b32731be56b5bac6de