Fix newline handling when adding nothing but newlines
This patch adds documentation and updates tests. The only actual
change is very minor: When no edit is made except for adding a
single linebreak, core's Diff algorithm represents this as a
( false → "" ) pair. The so called "AnnotatedHtmlDiffFormatter"
drops this information and represents both as "". This makes it
impossible to properly mark the empty line as "please keep".
Such lines are later dropped by the SplitConflictMerger.

The problematic line of code was introduced in Iedb9a0c, via
T250262.

Note this is not a problem when other edits are made the same
time. In this case the diff pair might be something like
( "A" → "B\n" ). This was never dropped. That's why users
don't see this issue often.

Bug: T284994
Change-Id: I70ec5548bbf711cab076a7201776ffa781099e6c