Update SplitConflictMerger tests to reflect reality better
This is done in addition to I70ec554. Instead of "assuming"
something we know if a line was empty before or emptied by
the user.

The merger was always able to deal with the necessary special
cases. The problem was that the merger didn't got all
information. There was no difference between a line that
didn't exist before (marked with "0"), and one that existed
but was empty (also "0", but should be "0,was-empty").

I70ec554 makes sure this information is not lost.

This is actually a bug. Luckily it's rather rare that an
empty line is presented as a separate chunk in a conflict.
Non-empty lines are never removed.

Change-Id: Ic09ddcfde34de84271eb3e82e168f5c2561ac9bf