Pass the actual page for rather than special page
So building the textbox needs a real page to do that and calling
the $this->getPageTitle() passes a special page instead which seems
incorrect.

So this patch actually passes the title of the page to build the
textbox on to the consumers.

Thanks to Daniel Kinzler for spotting that during reviewing my code
around TextboxBuilder::class.

Depends-On: Ia55251ee7f730636d6e85bf069734ff462119f0d
Change-Id: I32a0888e84e2fab4340dee349cfd51f057fb293c