Remove bogus Beta opt-out migration code
While this looks like a product change (the migration logic is
removed) it's really just a bugfix. This code never worked, and
never can.

Note this code is not executed anyway, at the moment. It only
makes a difference when a wiki's setup is changed from Beta to
default, and users start editing their preferences. This won't
happen for a while, if ever.

I noticed this is pretty much a re-upload of Ib2d6080, but it's
smaller and leaves an explanation why the code was removed.

Bug: T260967
Change-Id: Ic9f60036a3065bd499209c5d5eeb76ed6ae2ec8b