[selenium] Merge some talk page browser tests
Setting up the conflict is very expensive. Testing some of these
things independent form each other does not gain much.

Note that the order how some tests are executed now partly depends
on the outcome of the test before. This is simillar to what was
done in the `tourdialog` spec. It seems a good compromise between
keeping things kind of apart and avoid setup costs.

Bug: T282935
Change-Id: I56e6c9344d76208ac6f353e6e845ef4d4fc3aeff