Talk page special case doesn't make sense in self-conflicts
The special talk page interface is build with the assumption that you
can *not* edit the other user's post, but only choose if your post
should go before or after the other. This doesn't make sense when you
wrote both.

The normal conflict resolution interface might not be ideal either in
such a situation. But this is to be discussed later.

Note that patchset 3 of this patch contains a draft that disables
both interfaces in case of a self-conflict, not only the talk page
special case.

Bug: T301685
Change-Id: I4411842977e352b10fdb489833fa497e2b36bbd5