Handle complexity exception in ResolutionSuggester
This is a simple stability and reliability improvement that should be
uncontroversial. A Diff that's to complex can never be a talk page
conflict. That's simple by definition.

Notes:
* It might not even be possibly to run into this exception at this
point. But this is not obvious from the code, and maybe I'm wrong. It
can't hurt to have this safe-guard.
* This doesn't solve T248307. That's (mostly) about the other place(s)
where this exception can show up.

Bug: T248307
Change-Id: I2b02313b4c12e7539f2f06551dac34591d8b4777