Load ext.TwoColConflict.JSCheck on mobile domain
This module was flagged by warnings in December 2022 as a module
that is being added to the page via OutputPage::addModules() but
silently being ignored in mobile contexts.

What this JavaScript module does is only used for logging in
TwoColConflictHooks::doEditPageBeforeConflictDiff(). There is no
reason to exclude mobile at this point. The fact that the UI
currently doesn't load on mobile is part of the tracking via
shouldTwoColConflictBeShown().

Other modules are left untouched for now as we didn't see any
warnings in production logs. T250942 exists to continue working on
mobile support.

Bug: T324723
Change-Id: I09e3900c14872ce517de20a254e612850343ec6e