Improve test coverage for "copy text" special page link
The subpage part after the special page name was not covered by any
test, it seems.

This partly reverts I9285b72 where the $this->linkTarget property
was changed to be a Title. Turns out it's not that hard to change it
back to use the more narrow interface again. The proposed code relies
on Title::__toString(), but this should be fine as it's now covered
by a test.

Bug: T254440
Change-Id: I5a05ffc9fe6be808463d6ed600c1afe5f8c19856