Setup OOUI instead of using blank theme in tests
The `tearDown` for some of the tests is currently calling
`Theme::setSingleton()` to reset the defaut theme to `null` after
the tests - during the tests the theme is set to `BlankTheme`.
This is fine when the test suite is run on its own. Unfortunately
in the case that a previous test class runs `OutputPage::setupOOUI`
and a subsequent class also runs `OutputPage::setupOOUI`, the fact
that the Theme singleton has been set to `null` by this test suite
leaves OOUI in a semi-initialized state.

Remove calls to `setSingleton` from setup and teardown and use a
single call to `OutputPage::setupOOUI()` instead. The
`setSingleton` calls were introduced as an optimisation to reduce
the amount of HTML generated by the tests, but the optimisation
effect is marginal. Some tests even run faster with the
setup/teardown removed, some slower, but the difference (< 10ms)
is probably within the normal variance of test execution times.
`OutputPage::setupOOUI()` has its own singleton and will do
nothing in the case a previous test has already run the setup.
We don't clear this out in the teardown because
`OutputPage::resetOOUI()` just runs the setup again, and there
are calls to `setupOOUI()` all over the codebase.

Note that we can only make this replacement for integration
tests. In the Unit tests, the container is not available so
`OutputPage::setupOOUI()` would fail. For Unit tests we just
remove the `tearDown` so that the singleton theme doesn't get
reset to `null`.

Bug: T367385
Change-Id: Icb6901f4d5b69b5e4e459f248b18831ed7ee1132