Add notheme hack to make edit buttons visible in dark mode
This is meant to be temporary, but is a quick fix for the (really
bad) problem that the edit buttons are currently invisible in dark
mode. The proper fix is to make the diff and edit fields dark. But
this is quite a bit more complicated and needs much more time.

I'm also updating a few more tokens that are relatively easy to
replace. This could have been part of one of the previous patches,
but these are merged already.

Bug: T373231
Change-Id: I38a5460e17ab81e346df1abc9d4220e2aa8f0d2a